Relocate DAGDecomposition.*
