+++ torch/csrc/jit/passes/autostream.h
@@ -29,8 +29,8 @@ POSSIBILITY OF SUCH DAMAGE.
 
 #pragma once
 
-#include <torch/csrc/jit/ir.h>
-#include <torch/csrc/jit/script/module.h>
+#include <torch/csrc/jit/ir/ir.h>
+#include <torch/csrc/jit/api/module.h>
 
 namespace torch {
 namespace jit {
